Cleanup folder structure a bit
