+++ src/app/admin/groups/page.tsx
@@ -1,4 +1,4 @@
-import ManageGroups from "components/ManageGroups";
+import ManageGroups from "components/admin/ManageGroups";
 import OnlyRoles from "components/OnlyRoles";
 import PageWithHeader from "components/PageWithHeader";
 import React from "react";
