+++ src/app/admin/timetable/page.tsx
@@ -1,7 +1,7 @@
 import PageWithHeader from "components/PageWithHeader";
 import React from "react";
 import OnlyRoles from "components/OnlyRoles";
-import CreateTimetable from "components/CreateTimetable";
+import CreateTimetable from "components/admin/CreateTimetable";
 
 export const metadata = {
     title: "MyBPHS - Admin / Órarend",
