+++ src/app/admin/users/page.tsx
@@ -1,4 +1,4 @@
-import CreateUsers from "components/CreateUsers";
+import CreateUsers from "components/admin/CreateUsers";
 import OnlyRoles from "components/OnlyRoles";
 import PageWithHeader from "components/PageWithHeader";
 import React from "react";
