+++ src/components/admin/NFCInput.tsx
